requirements.txt
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
google-api-python-client==2.95.0
selenium==4.11.2
python-dotenv==1.0.0
schedule==1.2.0
pytz==2024.1 